Fix tests, clippy, FRB
